mkfs: fix integer overflow when -f option is used on a 32-bit machine
Fix integer overflow where

    mkfs.exfat -f ...

is used to format a device larger than 8GB on a 32-bit machine.

Change exfat_write_zero() to accept off_t instead of size_t. off_t is
assumed to be always 64-bit which is the almost always case with modern
libcs.

Fixes: #342

+++ include/libexfat.h
@@ -176,7 +176,7 @@ int exfat_get_blk_dev_info(struct exfat_user_input *ui,
 		struct exfat_blk_dev *bd);
 ssize_t exfat_read(int fd, void *buf, size_t size, off_t offset);
 ssize_t exfat_write(int fd, void *buf, size_t size, off_t offset);
-ssize_t exfat_write_zero(int fd, size_t size, off_t offset);
+int exfat_write_zero(int fd, off_t size, off_t offset);
 int exfat_discard_blocks(int fd, uint64_t start, uint64_t len);
 
 size_t exfat_utf16_len(const __le16 *str, size_t max_size);
